+++ CHANGELOG.md
@@ -5,12 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] 2025-04-29
+### Added 
+- Eddington number stats
+- Artist cut-over point stat
+- Most popular year/month stats
+- Longest Listening Streak stat
+- Average Plays per Active Day stat
+- Most Active Weekday stat
+- Peak Listening Hour stat
+
+### Changed
+- Cleaned up stats grouping to make it more human-friendly.
+
+### Fixed
+- Fixed jumping of the loading screen on mobile (for real this time)
+
 ## [1.7.1] 2025-04-29
 ### Changed
 - General formating changes.  
 
 ### Fixed
-- Fixed jumping of loading screen on mobile.
+- Fixed jumping of the loading screen on mobile.
 
 ## [1.7.0] 2025-04-29
 ### Added